requirements.txt
torch==2.2.0+cpu
torchvision==0.17.0+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html
ftfy
regex
tqdm
fastapi
uvicorn
pillow
numpy<2
python-multipart